+++ ceammc/CHANGELOG.md
@@ -9,6 +9,7 @@
   - @clip property added to system.cursor to clip cursor XY coords
   - @norm property added to ui.env to do envelope value normalization on input
   - @value property added to seq.counter
+  - @dev property added to flow.space: to set deviation
 - new methods:
   - [polltime( method added to system.cursor to change global polltime
 - misc: